// ninja-profiles/src/components/types.ts
export interface Ninja {
  email: string;
  gitHub: string | null;
  highlighted: boolean;
  imagePortraitUrl: string;
  imageWallOfLeetUrl: string;
  linkedIn: string | null;
  mainText: string;
  manager: string;
  name: string;
  office: string;
  orgUnit: string;
  phoneNumber: string;
  published: boolean;
  stackOverflow: string | null;
  twitter: string | null;
}